change custom convert method to static in CustomConverter.java
