--- src/main/java/run/ikaros/server/custom/GroupVersionKindName.java
@@ -1,4 +1,0 @@
-package run.ikaros.server.custom;
-
-public record GroupVersionKindName(String group, String version, String kind, String name) {
-}
